services/gateway/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "python-multipart>=0.0.20",
    "bcrypt>=4.0",
    "sqlalchemy>=2.0",
    "python-dotenv>=1.0",
    "itsdangerous>=2.1",
    "green-to-red",
    "yt-bulk-dl",
    "edl-to-archive",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gateway"]

[tool.uv.sources]
green-to-red = { workspace = true }
yt-bulk-dl = { workspace = true }
edl-to-archive = { workspace = true }
